feat(zshrc-manager): share one health report and slim Health Check to issues
