+++ extensions/zshrc-manager/src/health-check.tsx
@@ -8,29 +8,13 @@
  */
 
 import { Action, ActionPanel, List, Icon, Color, showToast, Toast } from "@raycast/api";
-import { useState, useMemo, useEffect } from "react";
 import { useZshrcLoader } from "./hooks/useZshrcLoader";
-import { calculateStatistics } from "./utils/statistics";
+import { useHealthReport } from "./hooks/useHealthReport";
 import { getZshrcPath } from "./lib/zsh";
-import {
-  detectDuplicates,
-  detectBrokenSources,
-  type DuplicateResult,
-  type BrokenSourceResult,
-} from "./utils/validation";
 import { MODERN_COLORS } from "./constants";
 
 interface HealthCheckProps {
   searchBarAccessory?: React.ReactElement;
-}
-
-interface HealthIssue {
-  id: string;
-  severity: "error" | "warning" | "info";
-  title: string;
-  description: string;
-  category: "duplicates" | "broken" | "organization";
-  details?: string;
 }
 
 /**
@@ -48,160 +32,11 @@ function getSeverityStyle(severity: "error" | "warning" | "info"): { icon: Icon;
 }
 
 /**
- * Creates health issues from validation results
- */
-function createHealthIssues(
-  duplicateAliases: DuplicateResult,
-  duplicateExports: DuplicateResult,
-  brokenSources: BrokenSourceResult,
-): HealthIssue[] {
-  const issues: HealthIssue[] = [];
-
-  // Duplicate aliases
-  duplicateAliases.duplicates.forEach((dup) => {
-    issues.push({
-      id: `dup-alias-${dup.name}`,
-      severity: "warning",
-      title: `Duplicate alias: ${dup.name}`,
-      description: `Defined ${dup.count} times in: ${dup.sections.join(", ")}`,
-      category: "duplicates",
-      details: `The alias "${dup.name}" is defined ${dup.count} times. Only the last definition will be active.`,
-    });
-  });
-
-  // Duplicate exports
-  duplicateExports.duplicates.forEach((dup) => {
-    issues.push({
-      id: `dup-export-${dup.name}`,
-      severity: "warning",
-      title: `Duplicate export: ${dup.name}`,
-      description: `Defined ${dup.count} times in: ${dup.sections.join(", ")}`,
-      category: "duplicates",
-      details: `The export "${dup.name}" is defined ${dup.count} times. Only the last value will be used.`,
-    });
-  });
-
-  // Broken sources
-  brokenSources.brokenSources.forEach((broken) => {
-    issues.push({
-      id: `broken-${broken.path}`,
-      severity: "error",
-      title: `Missing source file`,
-      description: broken.path,
-      category: "broken",
-      details: `The file "${broken.expandedPath}" does not exist or is not readable. This will cause an error when the shell starts.`,
-    });
-  });
-
-  return issues;
-}
-
-/**
- * Calculates overall health score
- */
-function calculateHealthScore(issues: HealthIssue[]): { score: number; label: string; color: string } {
-  const errorCount = issues.filter((i) => i.severity === "error").length;
-  const warningCount = issues.filter((i) => i.severity === "warning").length;
-
-  // Deduct points for issues
-  let score = 100;
-  score -= errorCount * 20;
-  score -= warningCount * 5;
-  score = Math.max(0, score);
-
-  let label: string;
-  let color: string;
-
-  if (score >= 90) {
-    label = "Excellent";
-    color = MODERN_COLORS.success;
-  } else if (score >= 70) {
-    label = "Good";
-    color = MODERN_COLORS.primary;
-  } else if (score >= 50) {
-    label = "Fair";
-    color = MODERN_COLORS.warning;
-  } else {
-    label = "Needs Attention";
-    color = MODERN_COLORS.error;
-  }
-
-  return { score, label, color };
-}
-
-/**
  * Health check dashboard component
  */
 export default function HealthCheck({ searchBarAccessory }: HealthCheckProps) {
   const { sections, isLoading, refresh } = useZshrcLoader("Health Check");
-  const [brokenSources, setBrokenSources] = useState<BrokenSourceResult | null>(null);
-  const [isCheckingBroken, setIsCheckingBroken] = useState(false);
-
-  const stats = useMemo(() => (sections.length > 0 ? calculateStatistics(sections) : null), [sections]);
-
-  // Synchronous checks - preserve original section context from parsed data
-  const duplicateAliases = useMemo(() => {
-    if (!stats) return { duplicates: [], totalDuplicates: 0 };
-    // Use section info from the parsed sections if available
-    const aliasesWithSection = sections.flatMap((section) =>
-      stats.aliases
-        .filter((a) => section.content.includes(`alias ${a.name}=`))
-        .map((a) => ({ ...a, section: section.label })),
-    );
-    return detectDuplicates(
-      aliasesWithSection.length > 0 ? aliasesWithSection : stats.aliases.map((a) => ({ ...a, section: "zshrc" })),
-      "name",
-    );
-  }, [stats, sections]);
-
-  const duplicateExports = useMemo(() => {
-    if (!stats) return { duplicates: [], totalDuplicates: 0 };
-    // Use section info from the parsed sections if available
-    const exportsWithSection = sections.flatMap((section) =>
-      stats.exports
-        .filter((e) => section.content.includes(`export ${e.variable}=`) || section.content.includes(`${e.variable}=`))
-        .map((e) => ({ ...e, section: section.label })),
-    );
-    return detectDuplicates(
-      exportsWithSection.length > 0 ? exportsWithSection : stats.exports.map((e) => ({ ...e, section: "zshrc" })),
-      "variable",
-    );
-  }, [stats, sections]);
-
-  // Async check for broken sources
-  useEffect(() => {
-    async function checkBrokenSources() {
-      if (!stats || stats.sources.length === 0) {
-        setBrokenSources({ brokenSources: [], totalBroken: 0 });
-        return;
-      }
-
-      setIsCheckingBroken(true);
-      try {
-        const result = await detectBrokenSources(stats.sources.map((s) => ({ path: s.path })));
-        setBrokenSources(result);
-      } catch (error) {
-        // Log the error for debugging purposes
-        console.error("Failed to detect broken sources:", error);
-        setBrokenSources({ brokenSources: [], totalBroken: 0 });
-      } finally {
-        setIsCheckingBroken(false);
-      }
-    }
-
-    checkBrokenSources();
-  }, [stats]);
-
-  const issues = useMemo(() => {
-    if (!brokenSources) return [];
-    return createHealthIssues(duplicateAliases, duplicateExports, brokenSources);
-  }, [duplicateAliases, duplicateExports, brokenSources]);
-
-  const healthScore = useMemo(() => calculateHealthScore(issues), [issues]);
-
-  const errorIssues = issues.filter((i) => i.severity === "error");
-  const warningIssues = issues.filter((i) => i.severity === "warning");
-  const infoIssues = issues.filter((i) => i.severity === "info");
+  const { issues, errorIssues, warningIssues, infoIssues, isChecking } = useHealthReport(sections);
 
   const handleRefresh = async () => {
     refresh();
@@ -216,7 +51,7 @@ export default function HealthCheck({ searchBarAccessory }: HealthCheckProps) {
       navigationTitle="Health Check"
       searchBarPlaceholder="Filter Issues..."
       searchBarAccessory={searchBarAccessory as List.Props["searchBarAccessory"]}
-      isLoading={isLoading || isCheckingBroken}
+      isLoading={isLoading || isChecking}
       isShowingDetail
       actions={
         <ActionPanel>
@@ -230,43 +65,8 @@ export default function HealthCheck({ searchBarAccessory }: HealthCheckProps) {
         </ActionPanel>
       }
     >
-      {/* Health Score Summary */}
-      <List.Section title="Summary">
-        <List.Item
-          title="Health Score"
-          subtitle={`${healthScore.score}/100 - ${healthScore.label}`}
-          icon={{ source: Icon.Heartbeat, tintColor: healthScore.color }}
-          accessories={[
-            { text: `${errorIssues.length} errors`, icon: errorIssues.length > 0 ? Icon.XMarkCircle : undefined },
-            {
-              text: `${warningIssues.length} warnings`,
-              icon: warningIssues.length > 0 ? Icon.ExclamationMark : undefined,
-            },
-          ]}
-          detail={
-            <List.Item.Detail
-              markdown={`
-# Health Score: ${healthScore.score}/100
-
-**Status:** ${healthScore.label}
-
-## Summary
-- 🔴 **Errors:** ${errorIssues.length}
-- 🟡 **Warnings:** ${warningIssues.length}
-- 🔵 **Info:** ${infoIssues.length}
-
-## Statistics
-- **Total Aliases:** ${stats?.aliases.length || 0}
-- **Total Exports:** ${stats?.exports.length || 0}
-- **Total Sources:** ${stats?.sources.length || 0}
-
-${issues.length === 0 ? "✅ No issues found! Your configuration looks healthy." : ""}
-              `}
-            />
-          }
-        />
-      </List.Section>
-
+      {/* The score/statistics summary lives on the home surface's Health
+          row; this view is the issues themselves. */}
       {/* Error Issues */}
       {errorIssues.length > 0 && (
         <List.Section title="Errors" subtitle={`${errorIssues.length} issues require attention`}>
@@ -276,7 +76,6 @@ ${issues.length === 0 ? "✅ No issues found! Your configuration looks healthy."
               <List.Item
                 key={issue.id}
                 title={issue.title}
-                subtitle={issue.description}
                 icon={{ source: style.icon, tintColor: style.color }}
                 accessories={[{ tag: { value: issue.category, color: Color.Red } }]}
                 detail={
@@ -324,7 +123,6 @@ Fix this issue to prevent errors when starting your shell.
               <List.Item
                 key={issue.id}
                 title={issue.title}
-                subtitle={issue.description}
                 icon={{ source: style.icon, tintColor: style.color }}
                 accessories={[{ tag: { value: issue.category, color: Color.Orange } }]}
                 detail={
@@ -372,7 +170,6 @@ Consider reviewing this configuration to avoid confusion.
               <List.Item
                 key={issue.id}
                 title={issue.title}
-                subtitle={issue.description}
                 icon={{ source: style.icon, tintColor: style.color }}
                 accessories={[{ tag: { value: issue.category, color: Color.Blue } }]}
                 detail={
@@ -412,11 +209,10 @@ This is informational and may be intentional.
       )}
 
       {/* No Issues */}
-      {issues.length === 0 && !isLoading && !isCheckingBroken && (
+      {issues.length === 0 && !isLoading && !isChecking && (
         <List.Section title="No Issues Found">
           <List.Item
             title="Your configuration looks healthy!"
-            subtitle="No duplicates, conflicts, or broken sources detected"
             icon={{ source: Icon.CheckCircle, tintColor: MODERN_COLORS.success }}
             detail={
               <List.Item.Detail
